[PloobsEngine] QuadTerrain working with Deferred [PloobsEngine] Fixed Some precision issues in some Shaders

+++ PloobsEngine/PloobsEngineContent/Effects/Deferred/InstanceRenderGBuffer.fx
@@ -47,10 +47,10 @@ VertexShaderOutput VertexShaderFunction(VertexShaderInput input, float4x4 instan
 
 struct PixelShaderOutput
 {
-    half4 Color : COLOR0;
-    half4 Normal : COLOR1;
-    half4 Depth : COLOR2;
-    half4 Extra1 : COLOR3;
+    float4 Color : COLOR0;
+    float4 Normal : COLOR1;
+    float4 Depth : COLOR2;
+    float4 Extra1 : COLOR3;
 };
 
 PixelShaderOutput PixelShaderFunctionTexture(VertexShaderOutput input)
